gateway/requirements.txt
fastapi==0.115.2
gunicorn==23.0.0
httpx==0.27.2
prometheus-fastapi-instrumentator==6.1.0
pydantic-settings==2.5.2
PyYAML==6.0.2
slowapi==0.1.9
uvicorn[standard]==0.30.6
redis